Move data and display into separate functions split between scripts and the new domUpdate file
